main: Move payload related functions to payloads module
Add the payloads module which contains all payload-related functions.
